1.优化推送退单贝泰妮接口

+++ nascent-bridge/src/main/java/com/me/nascent/modules/reorder/entity/ReFund.java
@@ -93,7 +93,7 @@ public class ReFund {
     private String refuseReason;
 
     @TableField("shopId")
-    private int shopId;
+    private long shopId;
 
     @TableField("sid")
     private String sid;
